Register WS command recorder/info early
